# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Employee attrition analytics and prediction API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface CountItem {
  label: string;
  count: number;
  percentage: number;
}

export interface DashboardGroup {
  label: string;
  yes: number;
  no: number;
  total: number;
  rate: number;
}

export interface IncomePoint {
  bucket: string;
  yes: number;
  no: number;
}

export interface Dashboard {
  total_employees: number;
  employees_left: number;
  employees_stayed: number;
  attrition_rate: number;
  target_distribution: CountItem[];
  by_department: DashboardGroup[];
  by_job_role: DashboardGroup[];
  by_overtime: DashboardGroup[];
  income_by_attrition: IncomePoint[];
}

export interface ModelComparison {
  model: string;
  accuracy: number;
  precision: number;
  recall: number;
  f1_score: number;
}

export interface FeatureImportance {
  feature: string;
  importance: number;
}

export interface ModelMetrics {
  model_used: string;
  accuracy: number;
  precision: number;
  recall: number;
  f1_score: number;
  confusion_matrix: number[][];
  comparison: ModelComparison[];
  feature_importance: FeatureImportance[];
  trained_at: string;
}

export type EmployeeInputBusinessTravel = typeof EmployeeInputBusinessTravel[keyof typeof EmployeeInputBusinessTravel];


export const EmployeeInputBusinessTravel = {
  'Non-Travel': 'Non-Travel',
  Travel_Rarely: 'Travel_Rarely',
  Travel_Frequently: 'Travel_Frequently',
} as const;

export type EmployeeInputDepartment = typeof EmployeeInputDepartment[keyof typeof EmployeeInputDepartment];


export const EmployeeInputDepartment = {
  Sales: 'Sales',
  'Research_&_Development': 'Research & Development',
  Human_Resources: 'Human Resources',
} as const;

export type EmployeeInputEducationField = typeof EmployeeInputEducationField[keyof typeof EmployeeInputEducationField];


export const EmployeeInputEducationField = {
  Life_Sciences: 'Life Sciences',
  Medical: 'Medical',
  Marketing: 'Marketing',
  Technical_Degree: 'Technical Degree',
  Human_Resources: 'Human Resources',
  Other: 'Other',
} as const;

export type EmployeeInputGender = typeof EmployeeInputGender[keyof typeof EmployeeInputGender];


export const EmployeeInputGender = {
  Female: 'Female',
  Male: 'Male',
} as const;

export type EmployeeInputJobRole = typeof EmployeeInputJobRole[keyof typeof EmployeeInputJobRole];


export const EmployeeInputJobRole = {
  Sales_Executive: 'Sales Executive',
  Research_Scientist: 'Research Scientist',
  Laboratory_Technician: 'Laboratory Technician',
  Manufacturing_Director: 'Manufacturing Director',
  Healthcare_Representative: 'Healthcare Representative',
  Manager: 'Manager',
  Sales_Representative: 'Sales Representative',
  Research_Director: 'Research Director',
  Human_Resources: 'Human Resources',
} as const;

export type EmployeeInputMaritalStatus = typeof EmployeeInputMaritalStatus[keyof typeof EmployeeInputMaritalStatus];


export const EmployeeInputMaritalStatus = {
  Single: 'Single',
  Married: 'Married',
  Divorced: 'Divorced',
} as const;

export type EmployeeInputOverTime = typeof EmployeeInputOverTime[keyof typeof EmployeeInputOverTime];


export const EmployeeInputOverTime = {
  Yes: 'Yes',
  No: 'No',
} as const;

export interface EmployeeInput {
  /**
     * @minimum 18
     * @maximum 75
     */
  Age: number;
  BusinessTravel: EmployeeInputBusinessTravel;
  Department: EmployeeInputDepartment;
  /**
     * @minimum 1
     * @maximum 100
     */
  DistanceFromHome: number;
  /**
     * @minimum 1
     * @maximum 5
     */
  Education: number;
  EducationField: EmployeeInputEducationField;
  /**
     * @minimum 1
     * @maximum 4
     */
  EnvironmentSatisfaction: number;
  Gender: EmployeeInputGender;
  /**
     * @minimum 1
     * @maximum 4
     */
  JobInvolvement: number;
  /**
     * @minimum 1
     * @maximum 5
     */
  JobLevel: number;
  JobRole: EmployeeInputJobRole;
  /**
     * @minimum 1
     * @maximum 4
     */
  JobSatisfaction: number;
  MaritalStatus: EmployeeInputMaritalStatus;
  /**
     * @minimum 100
     * @maximum 20000
     */
  MonthlyIncome: number;
  /**
     * @minimum 0
     * @maximum 20
     */
  NumCompaniesWorked: number;
  OverTime: EmployeeInputOverTime;
  /**
     * @minimum 10
     * @maximum 30
     */
  PercentSalaryHike: number;
  /**
     * @minimum 1
     * @maximum 5
     */
  PerformanceRating: number;
  /**
     * @minimum 1
     * @maximum 4
     */
  RelationshipSatisfaction: number;
  /**
     * @minimum 0
     * @maximum 3
     */
  StockOptionLevel: number;
  /**
     * @minimum 0
     * @maximum 50
     */
  TotalWorkingYears: number;
  /**
     * @minimum 0
     * @maximum 20
     */
  TrainingTimesLastYear: number;
  /**
     * @minimum 1
     * @maximum 4
     */
  WorkLifeBalance: number;
  /**
     * @minimum 0
     * @maximum 50
     */
  YearsAtCompany: number;
  /**
     * @minimum 0
     * @maximum 20
     */
  YearsInCurrentRole: number;
  /**
     * @minimum 0
     * @maximum 20
     */
  YearsSinceLastPromotion: number;
  /**
     * @minimum 0
     * @maximum 20
     */
  YearsWithCurrManager: number;
}

export type PredictionPrediction = typeof PredictionPrediction[keyof typeof PredictionPrediction];


export const PredictionPrediction = {
  Yes: 'Yes',
  No: 'No',
} as const;

export type PredictionRiskLevel = typeof PredictionRiskLevel[keyof typeof PredictionRiskLevel];


export const PredictionRiskLevel = {
  Low: 'Low',
  Medium: 'Medium',
  High: 'High',
} as const;

export interface Prediction {
  prediction: PredictionPrediction;
  /**
     * @minimum 0
     * @maximum 1
     */
  attrition_probability: number;
  risk_level: PredictionRiskLevel;
  model_used: string;
}