# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Employee attrition analytics and prediction API
 * OpenAPI spec version: 0.1.0
 */
import type { CountItem } from './countItem';
import type { DashboardGroup } from './dashboardGroup';
import type { IncomePoint } from './incomePoint';

export interface Dashboard {
  total_employees: number;
  employees_left: number;
  employees_stayed: number;
  attrition_rate: number;
  target_distribution: CountItem[];
  by_department: DashboardGroup[];
  by_job_role: DashboardGroup[];
  by_overtime: DashboardGroup[];
  income_by_attrition: IncomePoint[];
}