# lib/api-zod/src/generated/types/employeeInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Employee attrition analytics and prediction API
 * OpenAPI spec version: 0.1.0
 */
import type { EmployeeInputBusinessTravel } from './employeeInputBusinessTravel';
import type { EmployeeInputDepartment } from './employeeInputDepartment';
import type { EmployeeInputEducationField } from './employeeInputEducationField';
import type { EmployeeInputGender } from './employeeInputGender';
import type { EmployeeInputJobRole } from './employeeInputJobRole';
import type { EmployeeInputMaritalStatus } from './employeeInputMaritalStatus';
import type { EmployeeInputOverTime } from './employeeInputOverTime';

export interface EmployeeInput {
  /**
     * @minimum 18
     * @maximum 75
     */
  Age: number;
  BusinessTravel: EmployeeInputBusinessTravel;
  Department: EmployeeInputDepartment;
  /**
     * @minimum 1
     * @maximum 100
     */
  DistanceFromHome: number;
  /**
     * @minimum 1
     * @maximum 5
     */
  Education: number;
  EducationField: EmployeeInputEducationField;
  /**
     * @minimum 1
     * @maximum 4
     */
  EnvironmentSatisfaction: number;
  Gender: EmployeeInputGender;
  /**
     * @minimum 1
     * @maximum 4
     */
  JobInvolvement: number;
  /**
     * @minimum 1
     * @maximum 5
     */
  JobLevel: number;
  JobRole: EmployeeInputJobRole;
  /**
     * @minimum 1
     * @maximum 4
     */
  JobSatisfaction: number;
  MaritalStatus: EmployeeInputMaritalStatus;
  /**
     * @minimum 100
     * @maximum 20000
     */
  MonthlyIncome: number;
  /**
     * @minimum 0
     * @maximum 20
     */
  NumCompaniesWorked: number;
  OverTime: EmployeeInputOverTime;
  /**
     * @minimum 10
     * @maximum 30
     */
  PercentSalaryHike: number;
  /**
     * @minimum 1
     * @maximum 5
     */
  PerformanceRating: number;
  /**
     * @minimum 1
     * @maximum 4
     */
  RelationshipSatisfaction: number;
  /**
     * @minimum 0
     * @maximum 3
     */
  StockOptionLevel: number;
  /**
     * @minimum 0
     * @maximum 50
     */
  TotalWorkingYears: number;
  /**
     * @minimum 0
     * @maximum 20
     */
  TrainingTimesLastYear: number;
  /**
     * @minimum 1
     * @maximum 4
     */
  WorkLifeBalance: number;
  /**
     * @minimum 0
     * @maximum 50
     */
  YearsAtCompany: number;
  /**
     * @minimum 0
     * @maximum 20
     */
  YearsInCurrentRole: number;
  /**
     * @minimum 0
     * @maximum 20
     */
  YearsSinceLastPromotion: number;
  /**
     * @minimum 0
     * @maximum 20
     */
  YearsWithCurrManager: number;
}