# lib/api-zod/src/generated/types/featureImportance.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Employee attrition analytics and prediction API
 * OpenAPI spec version: 0.1.0
 */

export interface FeatureImportance {
  feature: string;
  importance: number;
}