# lib/api-zod/src/generated/types/modelComparison.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Employee attrition analytics and prediction API
 * OpenAPI spec version: 0.1.0
 */

export interface ModelComparison {
  model: string;
  accuracy: number;
  precision: number;
  recall: number;
  f1_score: number;
}