# lib/api-zod/src/generated/types/modelMetrics.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Employee attrition analytics and prediction API
 * OpenAPI spec version: 0.1.0
 */
import type { FeatureImportance } from './featureImportance';
import type { ModelComparison } from './modelComparison';

export interface ModelMetrics {
  model_used: string;
  accuracy: number;
  precision: number;
  recall: number;
  f1_score: number;
  confusion_matrix: number[][];
  comparison: ModelComparison[];
  feature_importance: FeatureImportance[];
  trained_at: string;
}