# lib/api-zod/src/generated/types/prediction.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Employee attrition analytics and prediction API
 * OpenAPI spec version: 0.1.0
 */
import type { PredictionPrediction } from './predictionPrediction';
import type { PredictionRiskLevel } from './predictionRiskLevel';

export interface Prediction {
  prediction: PredictionPrediction;
  /**
     * @minimum 0
     * @maximum 1
     */
  attrition_probability: number;
  risk_level: PredictionRiskLevel;
  model_used: string;
}